Only report a single value for solving time. #30
