+++ src/utils/version.h
@@ -16,19 +16,8 @@ All rights reserved (see LICENSE).
 #define DEV 0
 #define RC 1
 
-std::string get_version(){
-  std::string version = std::to_string(MAJOR)
-    + "." + std::to_string(MINOR)
-    + "." + std::to_string(PATCH);
-  if(DEV){
-    version += "-dev";
-  }
-  else{
-    if(RC){
-      version += "-rc." + std::to_string(RC);
-    }
-  }
-  return version;
-}
+#include <string>
+
+std::string get_version();
 
 #endif
